requirements/common.txt
Django==3.2.*
psycopg2-binary==2.9.1
djangorestframework>=3.11.2,<4
djangorestframework-simplejwt==4.6.0
django-cors-headers==3.5.0
